source/requirements.txt
Django==4.0.3
Pillow==9.1.0
djangorestframework==3.13.1
